FamilyBCDataset updates Numerous changes to CMake and configure scripts Updates to cgnsplot and cgnscheck
git-svn-id: http://svn.code.sf.net/p/cgns/code/cgns/release/3.1.4@445 59ee3124-ab72-4556-8a47-5b88dfbcc004

+++ src/cgnstools/calclib/vec.c
@@ -1727,7 +1727,7 @@ int n;
                 VECFLOAT *vec = op->op_val.vec;
                 if (!index)
                     index = n + 1;
-                if (--index < 0 || index >= op->len)
+                if (--index < 0 || index >= (int)op->len)
                     longjmp (env_ptr, VECERR_BADINDX);
                 push_var ((double)(vec[index]));
                 index = 0;
